+++ lua/kznllm-v3/basic_presets.lua
@@ -1,11 +1,11 @@
 local utils = require 'kznllm-v3.utils'
-local buffer_manager = (require 'kznllm-v3.buffers').buffer_manager
+local buffer_manager = (require 'kznllm-v3.buffer').buffer_manager
 local Path = require 'plenary.path'
 local api = vim.api
 
-local BaseProvider = require 'kznllm-v3.providers.base'
-local anthropic = require 'kznllm-v3.providers.anthropic'
-local openai = require 'kznllm-v3.providers.openai'
+local BaseProvider = require 'kznllm-v3.provider'
+local anthropic = require 'kznllm-v3.specs.anthropic'
+local openai = require 'kznllm-v3.specs.openai'
 
 local M = {}
