+++ lua/kznllm-v3/specs/openai.lua
